+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v0.1.2...HEAD)
+### Fixed
+- Error when attempting to translate stats with reminder strings. Translation for
+  those is currently not supported.
 
 ## [0.1.2](https://github.com/eps1lon/poe-mods/compare/v0.1.1...v0.1.2) (2017-20-03)
 ### Added